# src/main/resources/static/contracts/api.yml
openapi: 3.0.1
info:
  description: Word Frequency API
  version: 1.0.0
  title: Word Frequency
  contact:
    email: en.varg@gmail.com
  license:
    name: private
tags:
  - name: WordFrequency
    description: Count words in file.
paths:
  /api/freqency:
    get:
      tags:
        - Frequency
      summary: Get word frequency in file
      description: >-
          Method get file path and K-word and return how many time K-word in text
      operationId: getTopFrequency
      parameters:
        - name: file
          in: query
          required: true
          schema:
            type: string
          description: Full path to file
        - name: limit
          in: query
          required: true
          schema:
              type: integer
              default: 10
          description: Limit of result
        - name: ignoreCase
          in: query
          required: false
          schema:
              type: boolean
              default: true
          description: Ignore Case

      responses:
        '200':
          description: Response top word frequency
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrequencyResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2020-08-07T07:09:41.889+00:00'
                path: /api/frequency
                status: 400
                error: Bad Request
                message: 'Bad parameter file'
                requestId: "4121d29f-2949"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2020-08-07T07:09:41.889+00:00'
                path: /api/frequency
                status: 500
                error: Internal server error
                message: ''
                requestId: "4121d29f-2949"
    post:
      tags:
        - Frequency
      summary: Receive a file and count top word frequency in file
      description: >-
        Method get K-word and receive file and return how many time K-word in text
      operationId: topFrequency
      parameters:
        - name: limit
          in: query
          required: true
          description: K-word to search
          schema:
            type: integer
            default: 10
        - in: query
          required: false
          description: Ignore Case
          name: ignoreCase
          schema:
            type: boolean
            default: true
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Some file. text-plain format.
            encoding:
              file:
                contentType: text/plain; charset=utf-8
      responses:
        '200':
          description: Response top word frequency
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrequencyResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2020-08-07T07:09:41.889+00:00'
                path: /api/frequency
                status: 400
                error: Bad Request
                message: 'Bad parameter file'
                requestId: "4121d29f-2949"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2020-08-07T07:09:41.889+00:00'
                path: /api/frequency
                status: 500
                error: Internal server error
                message: ''
                requestId: "4121d29f-2949"
  /api/count:
    get:
      tags:
        - WordFrequency
      summary: Get word frequency in file
      description: >-
        Method get file path and K-word and return how many time K-word in text
      operationId: getFrequency
      parameters:
        - in: query
          required: true
          name: file
          schema:
            type: string
          description: Full path to file
        - in: query
          required: true
          name: word
          description: K-word to search
          schema:
            type: string
        - in: query
          required: false
          description: Ignore Case
          name: ignoreCase
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Response count word frequency
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrequencyResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2020-08-07T07:09:41.889+00:00'
                path: /api/frequency
                status: 400
                error: Bad Request
                message: 'Bad parameter file'
                requestId: "4121d29f-2949"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2020-08-07T07:09:41.889+00:00'
                path: /api/frequency
                status: 500
                error: Internal server error
                message: ''
                requestId: "4121d29f-2949"
    post:
      tags:
        - WordFrequency
      summary: Receive a file and count word frequency in file
      description: >-
        Method get K-word and receive file and return how many time K-word in text
      operationId: frequency
      parameters:
        - in: query
          required: true
          name: word
          description: K-word to search
          schema:
            type: string
        - in: query
          required: false
          description: Ignore Case
          name: ignoreCase
          schema:
            type: boolean
            default: true
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Some file. text-plain format.
            encoding:
              file:
                contentType: text/plain; charset=utf-8
      responses:
        '200':
          description: Response count word frequency
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrequencyResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2020-08-07T07:09:41.889+00:00'
                path: /api/frequency
                status: 400
                error: Bad Request
                message: 'Bad parameter file'
                requestId: "4121d29f-2949"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2020-08-07T07:09:41.889+00:00'
                path: /api/frequency
                status: 500
                error: Internal server error
                message: ''
                requestId: "4121d29f-2949"
components:
  schemas:
    FrequencyResponse:
      type: object
      description: Model word frequency.
      properties:
        fileName:
          type: string
          example: '/home/app/some-file.txt'
          description: Full path and filename
        words:
          type: array
          items:
            $ref: '#/components/schemas/WordFrequency'
    WordFrequency:
      type: object
      description: Model word frequency.
      properties:
        word:
          type: string
          example: 'awesome'
          description: word in text
        frequency:
          type: integer
          example: 10
          description: How many time word found in text.
    ErrorResponse:
      type: object
      description: Model error.
      properties:
        timestamp:
          type: string
          example: '2020-08-07T07:09:41.889+00:00'
          description: Time of request
        path:
          type: string
          example: /api/frequency
          description: Request path
        status:
          type: integer
          example: 400
          description: Response status
        error:
          type: string
          example: Bad Request
          description: Response error
        message:
          type: string
          example: 'Bad parameter file'
          description: Response message
        requestId:
          type: string
          example: "4121d29f-2949"
          description: Request id